Dynamic Environment Introduction
The environment file is critical to the generation of certificates and
many other parts of the rpm. The values being hard coded make it
difficult to start up a new system with the rpm file without having to
hunt down the differnt places that the domain names need to be changed.
This introduces a change to use the system short and dns name to parse
out the name of the system for certs and use in services.

Signed-off-by: Travis Powell <trpowell@lanl.gov>